+++ src/RcsbFv3D/RcsbFv3DSequenceIdentity.tsx
@@ -19,15 +19,16 @@ import {RcsbFv3DCssConfig} from "./RcsbFv3DComponent";
 import {MolstarAlignmentLoader} from "../RcsbFvStructure/StructureUtils/MolstarAlignmentLoader";
 import {MsaBehaviourObserver} from "../RcsbFvStructure/StructureViewerBehaviour/MsaBehaviour";
 import {
-    SequenceIdentityPfvManagerFactory
-} from "../RcsbFvSequence/SequenceViews/RcsbView/PfvManagerFactoryImplementation/SequenceIdentityPfvManagerFactory";
-import {
     PolymerEntityInstanceInterface
 } from "@rcsb/rcsb-saguaro-app/build/dist/RcsbCollectTools/DataCollectors/PolymerEntityInstancesCollector";
 import {SearchQuery} from "@rcsb/rcsb-api-tools/build/RcsbSearch/Types/SearchQueryInterface";
 import {HelpLinkComponent} from "../RcsbFvSequence/SequenceViews/RcsbView/Components/HelpLinkComponent";
 import {AlignmentResponse} from "@rcsb/rcsb-api-tools/build/RcsbGraphQL/Types/Borrego/GqlTypes";
 import {DataContainer} from "../Utils/DataContainer";
+import {
+    MsaPfvManagerFactory, MsaPfvManagerInterface
+} from "../RcsbFvSequence/SequenceViews/RcsbView/PfvManagerFactoryImplementation/MsaPfvManagerFactory";
+import {buildSequenceIdentityAlignmentFv} from "@rcsb/rcsb-saguaro-app";
 
 export interface RcsbFv3DSequenceIdentityInterface  {
     elementId?: string;
@@ -42,7 +43,12 @@ export interface RcsbFv3DSequenceIdentityInterface  {
     cssConfig?: RcsbFv3DCssConfig;
 }
 
-export class RcsbFv3DSequenceIdentity extends RcsbFv3DAbstract<{groupId:string; query?: SearchQuery;alignmentResponseContainer: DataContainer<AlignmentResponse>;},LoadMolstarInterface|undefined,{viewerElement:string|HTMLElement,viewerProps:Partial<ViewerProps>},{context:any,module:RcsbFvModulePublicInterface}> {
+export class RcsbFv3DSequenceIdentity extends RcsbFv3DAbstract<
+        MsaPfvManagerInterface,
+        LoadMolstarInterface|undefined,
+        {viewerElement:string|HTMLElement; viewerProps:Partial<ViewerProps>;},
+        {context:{id:string}; module:RcsbFvModulePublicInterface;}
+    > {
     constructor(params:RcsbFv3DSequenceIdentityInterface){
         const elementId: string = params.elementId ?? uniqid("RcsbFv3D_");
         const alignmentResponseContainer:DataContainer<AlignmentResponse> = new DataContainer<AlignmentResponse>();
@@ -56,13 +62,14 @@ export class RcsbFv3DSequenceIdentity extends RcsbFv3DAbstract<{groupId:string; 
                     rcsbId: params.config.groupId,
                     additionalConfig: params.additionalConfig,
                     pfvParams:{
-                        groupId:params.config.groupId,
-                        query:params.config.query,
+                        id: params.config.groupId,
+                        query: params.config.query,
+                        buildMsaAlignmentFv: buildSequenceIdentityAlignmentFv,
                         alignmentResponseContainer
                     },
                     buildPfvOnMount: true,
-                    pfvManagerFactory: new SequenceIdentityPfvManagerFactory<LoadMolstarInterface>(),
-                    callbackManagerFactory: new MsaCallbackManagerFactory<LoadMolstarInterface, {context:{groupId:string} & Partial<PolymerEntityInstanceInterface>}>({
+                    pfvManagerFactory: new MsaPfvManagerFactory<LoadMolstarInterface>(),
+                    callbackManagerFactory: new MsaCallbackManagerFactory<LoadMolstarInterface, {context:{id:string} & Partial<PolymerEntityInstanceInterface>}>({
                         pluginLoadParamsDefinition,
                         alignmentResponseContainer
                     }),